Split some things to different files, change mem alloc
