+++ src/tree/destroy_tree.c
@@ -1,0 +1,6 @@
+#include <rt_tree.h>
+
+void	destroy_tree(t_rttree *tree)
+{
+
+}
